# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Family Budget API for Chege & Lydiah
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /joint-account/deposit:
    post:
      operationId: createDeposit
      tags: [joint-account]
      summary: Deposit money into the joint account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JointAccountTransaction"
  /joint-account/disbursement:
    post:
      operationId: createDisbursement
      tags: [joint-account]
      summary: Disburse money from the joint account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisbursementInput"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JointAccountTransaction"
  /joint-account/{id}:
    delete:
      operationId: deleteJointAccountTransaction
      tags: [joint-account]
      summary: Delete a joint account transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /members:
    get:
      operationId: getMembers
      tags: [members]
      summary: List all members with access to this app
      responses:
        "200":
          description: Members list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: addMember
      tags: [members]
      summary: Add a new member by Replit user ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberInput"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "400":
          description: Invalid input or member already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /members/{userId}:
    delete:
      operationId: removeMember
      tags: [members]
      summary: Remove a member
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Member:
      type: object
      properties:
        userId:
          type: string
        userName:
          type: ["string", "null"]
        addedAt:
          type: string
          format: date-time
      required:
        - userId
        - addedAt
    AddMemberInput:
      type: object
      properties:
        userId:
          type: string
      required:
        - userId
...
    JointAccountTransaction:
      type: object
      properties:
        id:
          type: number
        type:
          type: string
          description: deposit or disbursement
        amount:
          type: number
        description:
          type: string
        madeById:
          type: string
          nullable: true
        madeByName:
          type: string
          nullable: true
        expenseCategory:
          type: string
          nullable: true
          description: Expense category this disbursement covers (optional)
        date:
          type: string
          format: date
        createdAt:
          type: string
      required:
        - id
        - type
        - amount
        - description
        - date
        - createdAt
...
    DepositInput:
      type: object
      properties:
        amount:
          type: number
        description:
          type: string
        date:
          type: string
          format: date
        madeById:
          type: string
          description: Defaults to the logged-in user
      required:
        - amount
        - description
        - date
    DisbursementInput:
      type: object
      properties:
        amount:
          type: number
        description:
          type: string
        date:
          type: string
          format: date
        expenseCategory:
          type: string
          description: Optional expense category this disbursement is paying for
      required:
        - amount
        - description
        - date
...